+++ src/app/common/admin.module.ts
@@ -11,5 +11,3 @@ import { AdminComponent } from './admin/admin.component';
 	providers: []
 })
 export class AdminModule {}
-export { AdminComponent } from './admin/admin.component';
-export { AdminTopic, AdminTopics } from './admin/admin-topic.model';
