+++ src/app/core/admin/cache-entries/cache-entries.component.ts
@@ -6,14 +6,15 @@ import { BsModalRef, BsModalService } from 'ngx-bootstrap/modal';
 import { Observable } from 'rxjs';
 import { filter, first, switchMap } from 'rxjs/operators';
 
-import { ModalAction, ModalService } from '../../../common/modal.module';
+import { ModalAction } from '../../../common/modal/modal.model';
+import { ModalService } from '../../../common/modal/modal.service';
 import {
 	PagingOptions,
 	PagingResults,
 	SortableTableHeader,
 	SortDirection
 } from '../../../common/paging.module';
-import { SystemAlertService } from '../../../common/system-alert.module';
+import { SystemAlertService } from '../../../common/system-alert/system-alert.service';
 import { AsyTableDataSource } from '../../../common/table/asy-table-data-source';
 import { CacheEntriesService } from './cache-entries.service';
 import { CacheEntryModalComponent } from './cache-entry-modal.component';
