+++ src/app/core/admin/end-user-agreement/admin-edit-eua.component.ts
@@ -4,7 +4,7 @@ import { ActivatedRoute, Params, Router } from '@angular/router';
 import { untilDestroyed, UntilDestroy } from '@ngneat/until-destroy';
 import { switchMap } from 'rxjs';
 
-import { ModalService } from '../../../common/modal.module';
+import { ModalService } from '../../../common/modal/modal.service';
 import { EndUserAgreement } from './eua.model';
 import { EuaService } from './eua.service';
 import { ManageEuaComponent } from './manage-eua.component';
