+++ src/app/core/admin/messages/manage-message.component.ts
@@ -5,9 +5,9 @@ import { Router } from '@angular/router';
 import { untilDestroyed } from '@ngneat/until-destroy';
 import { Observable } from 'rxjs';
 import { first } from 'rxjs/operators';
-import { ModalService } from 'src/app/common/modal.module';
-import { SystemAlertService } from 'src/app/common/system-alert.module';
 
+import { ModalService } from '../../../common/modal/modal.service';
+import { SystemAlertService } from '../../../common/system-alert/system-alert.service';
 import { ConfigService } from '../../config.service';
 import { Message } from '../../messages/message.class';
