+++ src/app/core/admin/user-management/list-users/admin-list-users.component.ts
@@ -9,7 +9,7 @@ import { catchError, filter, first, switchMap } from 'rxjs/operators';
 import { ModalAction } from '../../../../common/modal/modal.model';
 import { ModalService } from '../../../../common/modal/modal.service';
 import { PagingOptions, PagingResults, SortDirection } from '../../../../common/paging.module';
-import { SystemAlertService } from '../../../../common/system-alert.module';
+import { SystemAlertService } from '../../../../common/system-alert/system-alert.service';
 import { AsyTableDataSource } from '../../../../common/table/asy-table-data-source';
 import { AsyFilterDirective } from '../../../../common/table/filter/asy-filter.directive';
 import { Role } from '../../../auth/role.model';
