+++ src/app/core/feedback/feedback-flyout/feedback-flyout.spec.ts
@@ -6,8 +6,8 @@ import { Router } from '@angular/router';
 
 import { NgSelectModule } from '@ng-select/ng-select';
 import { AsyncSubject } from 'rxjs';
-import { FlyoutComponent } from 'src/app/common/flyout/flyout.component';
 
+import { FlyoutComponent } from '../../../common/flyout/flyout.component';
 import { ConfigService } from '../../config.service';
 import { FeedbackService } from '../feedback.service';
 import { FeedbackFlyoutComponent } from './feedback-flyout.component';
