+++ src/app/core/teams/list-teams/list-teams.component.ts
@@ -4,8 +4,8 @@ import cloneDeep from 'lodash/cloneDeep';
 import { Observable } from 'rxjs';
 
 import { PagingOptions, PagingResults } from '../../../common/paging.module';
-import { SystemAlertService } from '../../../common/system-alert.module';
-import { AsyTableDataSource } from '../../../common/table.module';
+import { SystemAlertService } from '../../../common/system-alert/system-alert.service';
+import { AsyTableDataSource } from '../../../common/table/asy-table-data-source';
 import { AuthorizationService } from '../../auth/authorization.service';
 import { Team } from '../team.model';
 import { TeamsService } from '../teams.service';
